[FIX] Moved from javascript to typescript
